app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anatoliy Pestich <aa@pestich.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
orjson = "^3.10.0"
pandas = "^2.2.2"
aiobotocore = "^2.12.3"
minio = "^7.2.5"
aiofiles = "^23.2.1"
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
